Move cmd APIs to internal/cmd/operator-sdk
