+++ internal/cmd/operator-sdk/bundle/cmd_test.go
