+++ internal/cmd/operator-sdk/bundle/internal/logger.go
