+++ internal/cmd/operator-sdk/bundle/internal/result.go
