+++ internal/cmd/operator-sdk/bundle/internal/result_test.go
