+++ internal/cmd/operator-sdk/cleanup/packagemanifests/packagemanifests.go
