+++ internal/cmd/operator-sdk/completion/bash.go
