+++ internal/cmd/operator-sdk/completion/bash_test.go
