+++ internal/cmd/operator-sdk/completion/cmd.go
