+++ internal/cmd/operator-sdk/completion/cmd_test.go
