+++ internal/cmd/operator-sdk/completion/completion_suite_test.go
