+++ internal/cmd/operator-sdk/completion/zsh.go
