+++ internal/cmd/operator-sdk/completion/zsh_test.go
