+++ internal/cmd/operator-sdk/generate/bundle/cmd.go
