+++ internal/cmd/operator-sdk/generate/internal/genutil.go
