+++ internal/cmd/operator-sdk/generate/kustomize/cmd.go
