+++ internal/cmd/operator-sdk/generate/kustomize/manifests.go
