+++ internal/cmd/operator-sdk/generate/packagemanifests/cmd.go
