+++ internal/cmd/operator-sdk/olm/cmd_test.go
