+++ internal/cmd/operator-sdk/olm/status_test.go
