+++ internal/cmd/operator-sdk/olm/uninstall.go
