+++ internal/cmd/operator-sdk/olm/uninstall_test.go
