+++ internal/cmd/operator-sdk/run/cmd_test.go
