+++ internal/cmd/operator-sdk/run/packagemanifests/packagemanifests_test.go
