+++ internal/cmd/operator-sdk/scorecard/cmd.go
