+++ internal/cmd/operator-sdk/scorecard/cmd_test.go
