+++ internal/cmd/operator-sdk/scorecard/scorecard_suite_test.go
