+++ internal/cmd/operator-sdk/version/cmd.go
