+++ internal/cmd/operator-sdk/version/cmd_test.go
