+++ internal/cmd/operator-sdk/version/version_suite_test.go
